Cypress tests working (except for some missing 44007 wind values)
