+++ cypress/integration/Platform/m01_spec.js
@@ -9,6 +9,8 @@ describe("Platfrom M01", () => {
     cy.contains("Regions").click()
     cy.contains("Gulf Of Maine").click()
 
+    cy.contains("Platforms in Gulf Of Maine")
+
     cy.contains("M01").click()
 
     cy.contains("Station M01")
@@ -29,14 +31,14 @@ describe("Platfrom M01", () => {
     cy.contains("Latest Conditions")
     cy.contains("Air Temperature -")
 
-    cy.get("[style='margin-top: 1rem;'] > :nth-child(2) > .row").children().its("length").should("be.gte", 4)
+    cy.get("#currentConditions > .col > .row").children().its("length").should("be.gte", 4)
   })
 
   xit("Shows wind plot", () => {
     cy.visit(platformUrl)
 
     cy.contains("Observations").click()
-    cy.get('[href="/platform/M01 - 44037/observations/wind"]').first().click()
+    cy.get('[href="/platform/M01%20-%2044037/observations/wind"]').first().click()
     cy.get("h4").contains("Wind")
     cy.get("svg.highcharts-root")
     cy.get("svg.highcharts-root").contains("Gust").click()
@@ -48,9 +50,9 @@ describe("Platfrom M01", () => {
   it("Shows wave forecast", () => {
     cy.visit(platformUrl)
 
-    cy.contains("Forecasts loading")
+    // cy.contains("Forecasts loading")
     cy.get("#forecast").click()
-    cy.get("[href='/platform/M01 - 44037/forecast/significant_wave_height']").click()
+    cy.get('[href="/platform/M01%20-%2044037/forecast/significant_wave_height"]').click()
     cy.get("h4").contains("Significant Wave Height Forecast")
 
     cy.get("svg.highcharts-root").contains("Feet")
