+++ src/Features/ERDDAP/Platform/Forecasts/dynamic.tsx
@@ -1,0 +1,3 @@
+import dynamic from "next/dynamic"
+
+export const Forecast = dynamic(() => import("./index"), { ssr: false })
