+++ src/Features/ERDDAP/hooks/index.ts
@@ -1,4 +1,12 @@
-export { usePlatforms, useForecast, useForecasts, getPlatforms, BUOY_BARN_PLATFORMS_KEY } from "./buoyBarn"
+export {
+  usePlatforms,
+  useForecast,
+  useForecasts,
+  getPlatforms,
+  BUOY_BARN_PLATFORMS_KEY,
+  BUOY_BARN_FORECAST_KEY,
+  getForecasts,
+} from "./buoyBarn"
 export { useDataset, useDatasets } from "./tabledap"
 export { UsePlatform, UsePlatforms } from "./BuoyBarnComponents"
 export { UseDataset, UseDatasets } from "./TableDAPComponents"
