+++ src/Shared/constants/index.ts
@@ -1,2 +1,2 @@
 export { regionList } from "./regions"
-export { paths } from "./routes"
+export { PATHS } from "./routes"
