+++ src/Shared/constants/routes.ts
@@ -1,17 +1,17 @@
 /**
  * URL paths for various pages
  */
-export const paths = {
+export const PATHS = {
   about: "/about",
   home: "/",
   map: "/map",
   platforms: {
-    all: "/platform/:id/observations/all",
-    forecast: "/platform/:id/forecast",
-    forecastType: "/platform/:id/forecast/:type",
-    observations: "/platform/:id/observations/:type",
-    observationsWind: "/platform/:id/observations/wind",
-    platform: "/platform/:id",
-    root: "/platform/"
-  }
+    all: "/platform/[id]/observations/all",
+    forecast: "/platform/[id]/forecast",
+    forecastType: "/platform/[id]/forecast/[type]",
+    observations: "/platform/[id]/observations/[type]",
+    observationsWind: "/platform/[id]/observations/wind",
+    platform: "/platform/[id]",
+    root: "/platform/",
+  },
 }
