+++ src/components/Head/index.tsx
@@ -8,5 +8,11 @@ interface HeadProps {
 export const Head: React.FunctionComponent<HeadProps> = ({ pageName }) => (
   <NextHead>
     <title>NERACOOS Mariners Dashboard{pageName ? " - " + pageName : null}</title>
+
+    <link
+      rel="shortcut icon"
+      href="http://neracoos.org/sites/neracoos.org/files/neracoosfavicon_0.ico"
+      type="image/vnd.microsoft.icon"
+    ></link>
   </NextHead>
 )
