+++ src/pages/platform/[id]/index.tsx
@@ -2,18 +2,36 @@ import { GetStaticPaths, GetStaticProps } from "next"
 import React from "react"
 import { QueryCache } from "react-query"
 import { dehydrate } from "react-query/hydration"
+import { Col, Row } from "reactstrap"
 
 import { PlatformLayout } from "components/Layout"
-import { ErddapCurrentPlatformConditions, BUOY_BARN_PLATFORMS_KEY, getPlatforms } from "Features/ERDDAP"
+import {
+  ErddapCurrentPlatformConditions,
+  BUOY_BARN_PLATFORMS_KEY,
+  getPlatforms,
+  BUOY_BARN_FORECAST_KEY,
+  getForecasts,
+} from "Features/ERDDAP"
 
 const CurrentConditions: React.FunctionComponent = () => {
-  return <PlatformLayout>{({ platform }) => <ErddapCurrentPlatformConditions platform={platform} />}</PlatformLayout>
+  return (
+    <PlatformLayout>
+      {({ platform }) => (
+        <Row id="currentConditions">
+          <Col>
+            <ErddapCurrentPlatformConditions platform={platform} />
+          </Col>
+        </Row>
+      )}
+    </PlatformLayout>
+  )
 }
 
 export const getStaticProps: GetStaticProps = async () => {
   const queryCache = new QueryCache()
 
   await queryCache.prefetchQuery(BUOY_BARN_PLATFORMS_KEY, getPlatforms)
+  await queryCache.prefetchQuery(BUOY_BARN_FORECAST_KEY, getForecasts)
 
   return {
     props: {
@@ -24,12 +42,8 @@ export const getStaticProps: GetStaticProps = async () => {
 }
 
 export const getStaticPaths: GetStaticPaths = async () => {
-  const platforms = await getPlatforms()
-
-  const paths = platforms.features.map((platform) => ({ params: { id: platform.id } }))
-
   return {
-    paths,
+    paths: [],
     fallback: true,
   }
 }
